set starting frame number of next acquisition for both jungfrau and e… (#27)
* set starting frame number of next acquisition for both jungfrau and eiger. firmware has not implemented a get, so workaround. tests included. frame number 0 not allowed due to Eiger. Eiger max frame is 48 bit, while jungfrau is 64 bit

* made argument of setstartingframenumber const

+++ slsSupportLib/src/DataSocket.cpp
@@ -65,7 +65,7 @@ int DataSocket::setReceiveTimeout(int us) {
 
 int DataSocket::sendData(const void *buffer, size_t size) {
     int dataSent = 0;
-    while (dataSent < size) {
+    while (dataSent < (int)size) {
         dataSent +=
             write(getSocketId(), reinterpret_cast<const char *>(buffer) + dataSent,
                   size - dataSent);
